Convert package for easier consumption
Modded the pkg format to be easier to consume
