+++ pkg/phantom/conversion.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 	"strings"
 	"time"
-	"phantom/pkg/socket/wire"
+	"github.com/breakcrypto/phantom/pkg/socket/wire"
 )
 
 func ConvertVersionStringToInt(str string) uint32 {
