+++ pkg/phantom/preamble.go
@@ -1,0 +1,78 @@
+/**
+*    Copyright (C) 2019-present C2CV Holdings, LLC.
+*
+*    This program is free software: you can redistribute it and/or modify
+*    it under the terms of the Server Side Public License, version 1,
+*    as published by C2CV Holdings, LLC.
+*
+*    This program is distributed in the hope that it will be useful,
+*    but WITHOUT ANY WARRANTY; without even the implied warranty of
+*    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+*    Server Side Public License for more details.
+*
+*    You should have received a copy of the Server Side Public License
+*    along with this program. If not, see
+*    <http://www.mongodb.com/licensing/server-side-public-license>.
+*
+*    As a special exception, the copyright holders give permission to link the
+*    code of portions of this program with the OpenSSL library under certain
+*    conditions as described in each individual source file and distribute
+*    linked combinations including the program with the OpenSSL library. You
+*    must comply with the Server Side Public License in all respects for
+*    all of the code used other than as permitted herein. If you modify file(s)
+*    with this exception, you may extend this exception to your version of the
+*    file(s), but you are not obligated to do so. If you do not wish to do so,
+*    delete this exception statement from your version. If you delete this
+*    exception statement from all source files in the program, then also delete
+*    it in the license file.
+*/
+
+package phantom
+
+import "fmt"
+
+func Preamble(version string) {
+	fmt.Println()
+	fmt.Println()
+	fmt.Println(`/**`)
+	fmt.Println(`*    Copyright (C) 2019-present C2CV Holdings, LLC.`)
+	fmt.Println(`*`)
+	fmt.Println(`*    This program is free software: you can redistribute it and/or modify`)
+	fmt.Println(`*    it under the terms of the Server Side Public License, version 1,`)
+	fmt.Println(`*    as published by C2CV Holdings, LLC.`)
+	fmt.Println(`*`)
+	fmt.Println(`*    This program is distributed in the hope that it will be useful,`)
+	fmt.Println(`*    but WITHOUT ANY WARRANTY; without even the implied warranty of`)
+	fmt.Println(`*    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the`)
+	fmt.Println(`*    Server Side Public License for more details.`)
+	fmt.Println(`*`)
+	fmt.Println(`*    You should have received a copy of the Server Side Public License`)
+	fmt.Println(`*    along with this program. If not, see`)
+	fmt.Println(`*    <http://www.mongodb.com/licensing/server-side-public-license>.`)
+	fmt.Println(`*`)
+	fmt.Println(`*    As a special exception, the copyright holders give permission to link the`)
+	fmt.Println(`*    code of portions of this program with the OpenSSL library under certain`)
+	fmt.Println(`*    conditions as described in each individual source file and distribute`)
+	fmt.Println(`*    linked combinations including the program with the OpenSSL library. You`)
+	fmt.Println(`*    must comply with the Server Side Public License in all respects for`)
+	fmt.Println(`*    all of the code used other than as permitted herein. If you modify file(s)`)
+	fmt.Println(`*    with this exception, you may extend this exception to your version of the`)
+	fmt.Println(`*    file(s), but you are not obligated to do so. If you do not wish to do so,`)
+	fmt.Println(`*    delete this exception statement from your version. If you delete this`)
+	fmt.Println(`*    exception statement from all source files in the program, then also delete`)
+	fmt.Println(`*    it in the license file.`)
+	fmt.Println(`*/`)
+	fmt.Println()
+	fmt.Printf("--- breakcrypto's phantoms ( v. %s ) ---\n", version)
+	fmt.Println()
+	fmt.Println("email: breakcrypto@gmail.com")
+	fmt.Println("twitter: @_breakcrypto")
+	fmt.Println("discord: breakcrypto#0011")
+	fmt.Println()
+	fmt.Println("Donation addresses:")
+	fmt.Println("BTC: 151HTde9NgwbMMbMmqqpJYruYRL4SLZg1S")
+	fmt.Println("LTC: LhBx1TUyp7wiYuMxjefAGUGZVzuHRtPBA7")
+	fmt.Println("DOGE: DBahutcjEAxfwQEW7kzft2y8dhZN2VtcG5")
+	fmt.Println()
+	fmt.Println()
+}
